requirements.txt
kafka-python==2.0.2
faker==19.3.1
psycopg2-binary==2.9.7
pandas==2.0.3
plotly==5.16.1
streamlit==1.26.0
sqlalchemy==2.0.20
